requirements.txt
streamlit
openai
pandas
rapidfuzz
easyocr
Pillow
